Add support for builder-hub messages
This commit adds support for VTEX/builder-hub messages, replacing
the `toggle` routes.

Check https://github.com/vtex/builder-hub/pull/280 for more information.
